feat: extract 2 routes and update test
